separate the interfaces
