+++ src/interfaces/Category.ts
@@ -1,0 +1,7 @@
+export interface Category {
+    id: number;
+    name: string;
+    image: string;
+    creationAt: string;
+    updatedAt: string;
+}
